+++ generators/_init-hybrid/templates/_gitignore
@@ -21,6 +21,4 @@ __pycache__/
 node_modules/
 /src/**/*.js
 /tests/**/*.js
-*.map
-*.css
 /.cache-loader
