+++ generators/_init-hybrid/templates/package.tmpl.json
@@ -7,11 +7,10 @@
     "dist"
   ],
   "scripts": {
-    "clean": "rimraf dist lib && npm run clean:compile",
-    "clean:compile": "rimraf src/**/*.map src/**/*.js src/**/*.d.ts tests/**/*.js tests/**/*.map tests/**/*.d.ts demo/**/*.js demo/**/*.map demo/**/*.d.ts",
+    "clean": "rimraf dist lib",
     "compile": "tsc",
     "compile:watch": "tsc -w",
-     "lint": "tslint -c tslint.json -p . 'src/**/*.ts?(x)' 'tests/**/*.ts?(x)'",    
+    "lint": "tslint -c tslint.json -p . 'src/**/*.ts?(x)' 'tests/**/*.ts?(x)'",    
     "check": "flake8 --exclude=.git,venv,deploy,docs,__pycache__,node_modules",
     "docs:web": "typedoc --options typedoc.json src/",
     "docs:python": "sphinx-apidoc -o docs -f ./<%-name.toLowerCase()%> && sphinx-build ./docs build/docs",
@@ -27,9 +26,7 @@
     "prebuild:web": "node -e \"process.exit(process.env.PHOVEA_SKIP_TESTS === undefined?1:0)\" || npm run test:web",
     "prebuild": "echo hybrid",
     "build:python": "rm -rf build/source && find . -name '*.pyc' -delete && node buildPython.js && cp -r ./<%-name.toLowerCase()%> build/source/",
-    "build:web": "npm run clean && npm run compile && npm run copy-assets && npm run copy-styles",
-    "copy-assets": "if [ -d src/assets ]; then shx --verbose cp -R src/assets/. dist/assets/; fi",
-    "copy-styles": "if [ -d src/scss ]; then shx --verbose cp -R src/scss/. dist/scss/; fi",    
+    "build:web": "npm run clean && npm run compile",
     "build": "npm run build:web && npm run build:python",
     "predist:web": "npm run build:web && npm run docs:web",
     "predist:python": "npm run build:python && npm run docs:python",
@@ -37,5 +34,18 @@
     "dist:web": "mkdir lib && cd dist && tar cvzf ../lib/<%-name%>.tar.gz *",
     "dist:python": "python setup.py sdist bdist_wheel",
     "dist": "npm run dist:web && npm run dist:python"
+  },
+  "devDependencies": {
+    "@types/jest": "~26.0.5",
+    "identity-obj-proxy": "~3.0.0",
+    "jest": "~26.1.0",
+    "jest-raw-loader": "~1.0.1",
+    "rimraf": "~3.0.2",
+    "shx": "~0.3.2",
+    "ts-jest": "~26.1.3",
+    "tslib": "~2.0.0",
+    "tslint": "~5.20.1",
+    "typedoc": "~0.17.8",
+    "typescript": "~3.9.7"
   }
 }
