+++ generators/_init-python/templates/_gitignore
@@ -2,7 +2,7 @@
 /.tscache
 /.idea
 /build/
-/dist/
+/lib/
 *.egg-info/
 *.egg
 *.py[cod]
