+++ generators/_init-web/templates/_gitignore
@@ -5,7 +5,5 @@
 node_modules/
 /src/**/*.js
 /tests/**/*.js
-*.map
-*.css
 *.log
 /.cache-loader
