+++ generators/_init-web/templates/package.tmpl.json
@@ -11,36 +11,33 @@
     "node": ">= 12.13"
   },
   "scripts": {
-    "posttest": "npm run lint",
-    "clean": "rimraf dist lib && npm run clean:compile",
-    "clean:compile": "rimraf src/**/*.map src/**/*.js src/**/*.d.ts tests/**/*.js tests/**/*.map tests/**/*.d.ts demo/**/*.js demo/**/*.map demo/**/*.d.ts",
-    "compile:watch": "tsc -w",
-    "build": "npm run clean && npm run compile && npm run copy-assets && npm run copy-styles",
-    "copy-assets": "if [ -d src/assets ]; then shx --verbose cp -R src/assets/. dist/assets/; fi",
-    "copy-styles": "if [ -d src/scss ]; then shx --verbose cp -R src/scss/. dist/scss/; fi",
-    "dist": "mkdir lib && cd dist && tar cvzf ../lib/<%-name%>.tar.gz *",
+    "clean": "rimraf dist lib",
     "compile": "tsc",
+    "compile:watch": "tsc -w",
     "lint": "tslint -c tslint.json -p . 'src/**/*.ts?(x)' 'tests/**/*.ts?(x)'",
     "docs": "typedoc --options typedoc.json src/",
-    "prebuild": "node -e \"process.exit(process.env.PHOVEA_SKIP_TESTS === undefined?1:0)\" || npm run test",
-    "test": "jest",
-    "test:watch": "jest --watch",
+    "test": "test ! $(find tests -name \"*.ts\") || jest --maxWorkers=1",
+    "posttest": "npm run lint",
+    "prebuild": "node -e \"process.exit(process.env.PHOVEA_SKIP_TESTS === undefined?1:0)\" || npm run test",    
+    "build": "npm run clean && npm run compile",
+    "predist": "npm run build && npm run docs",
+    "dist": "mkdir lib && cd dist && tar cvzf ../lib/<%-name%>.tar.gz *",
     "release:major": "npm version major && npm publish && git push --follow-tags",
     "release:minor": "npm version minor && npm publish && git push --follow-tags",
-    "release:patch": "npm version patch && npm publish && git push --follow-tags",
-    "predist": "npm run build && npm run docs"
+    "release:patch": "npm version patch && npm publish && git push --follow-tags"
+
   },
   "devDependencies": {
-    "@types/jest": "24.9.1",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "24.9.0",
-    "jest-raw-loader": "1.0.1",
-    "tslib": "~1.11.0",
-    "tslint": "5.9.1",
-    "ts-jest": "24.3.0",
-    "typedoc": "~0.16.10",
-    "typescript": "~3.8.2",
-    "rimraf": "^3.0.0",
-    "shx": "^0.3.2"
+    "@types/jest": "~26.0.5",
+    "identity-obj-proxy": "~3.0.0",
+    "jest": "~26.1.0",
+    "jest-raw-loader": "~1.0.1",
+    "rimraf": "~3.0.2",
+    "shx": "~0.3.2",
+    "ts-jest": "~26.1.3",
+    "tslib": "~2.0.0",
+    "tslint": "~5.20.1",
+    "typedoc": "~0.17.8",
+    "typescript": "~3.9.7"
   }
 }
