+++ generators/_init-web/templates/plain/tsd.d.ts
@@ -28,3 +28,4 @@ declare module "*.xml" {
   const content: string;
   export default content;
 }
+declare module "*.json";
