+++ generators/init-app-slib/templates/package.tmpl.json
