+++ generators/init-app/templates/sample_plain/src/app/App.ts
@@ -3,7 +3,7 @@
  */
 
 import * as d3 from 'd3';
-import {HELLO_WORLD} from './language';
+import {HELLO_WORLD} from '../language';
 
 /**
  * The main class for the App app
@@ -42,13 +42,12 @@ export class App {
     this.$node.select('.busy').classed('hidden', !isBusy);
   }
 
-}
-
-/**
- * Factory method to create a new app instance
- * @param parent
- * @returns {App}
- */
-export function create(parent:Element) {
-  return new App(parent);
+  /**
+   * Factory method to create a new app instance
+   * @param parent
+   * @returns {App}
+   */
+  static create(parent:Element) {
+    return new App(parent);
+  }
 }
