+++ generators/init-app/templates/sample_plain/src/app/index.ts
@@ -1,0 +1,1 @@
+export * from './App';
