+++ generators/init-app/templates/sample_plain/src/scss/_base.scss
