+++ generators/init-app/templates/sample_plain/src/scss/main.scss
@@ -1,0 +1,2 @@
+
+@import "./base";
