--- generators/init-app/templates/sample_plain/src/style.scss
@@ -1,2 +1,0 @@
-
-@import "styles/base";
