+++ generators/init-app/templates/sample_processed/src/index.template.ejs
@@ -1,34 +1,35 @@
 <!doctype html>
 <html>
+
 <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title><%= app %></title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width">
-    <link href="//fonts.googleapis.com/css?family=Yantramanav:700,500,400" rel="stylesheet" type="text/css">
-    <script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
-    <link href="./style.css?cache-bust=${Date.now()}" rel="stylesheet" type="text/css">
+  <meta charset="utf-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <title><%= app %></title>
+  <meta name="description" content="">
+  <meta name="viewport" content="width=device-width">
+  <link href="//fonts.googleapis.com/css?family=Yantramanav:700,500,400" rel="stylesheet" type="text/css">
+  <script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
 </head>
+
 <body>
 
-<div id="caleydoHeader"></div>
+  <div id="caleydoHeader"></div>
 
-<div id="app">
+  <div id="app">
     <h3 style="margin: 5px;">Loading &hellip;</h3>
     <div class="busy"></div>
-</div>
+  </div>
 
-<script>
-  (function() {
-    var isChrome = /Chrome/.test(navigator.userAgent) && /Google Inc/.test(navigator.vendor);
-    if (!isChrome) {
-      alert('This application was designed to be used in Google Chrome. Running the application in other browsers might result in performance issues or other misbehavior.');
-    }
-  })();
-</script>
-<script src="./common.js?cache-bust=${Date.now()}"></script>
-<script src="./app.js?cache-bust=${Date.now()}"<%-clientOnly ? ' data-phovea-offline="true"': ''%>></script>
+
+  <script>
+    (function () {
+      var isChrome = /Chrome/.test(navigator.userAgent) && /Google Inc/.test(navigator.vendor);
+      if (!isChrome) {
+        alert('This application was designed to be used in Google Chrome.\nRunning the application in other browsers might result in performance issues or other misbehavior.');
+      }
+    })();
+  </script>
 
 </body>
+
 </html>
