+++ generators/init-lib/templates/sample_plain/src/index.ts
@@ -1,14 +1,1 @@
-
-import {mixin} from 'phovea_core/src';
-
-export interface IHelloOptions {
-  name?: string;
-}
-
-export function hello(options?: IHelloOptions) {
-  //merge with default options
-  options = mixin({
-    name: 'World'
-  }, options);
-  return `Hello ${options.name}`;
-}
+export * from './hello';
